Added flash messages

+++ src/Controller/PostController.php
@@ -26,6 +26,7 @@ class PostController extends AbstractController
     public function index(PostRepository $postRepository): Response
     {
         return $this->render('post/index.html.twig', [
+            'tags' => $this->getDoctrine()->getManager()->getRepository(PostTags::class)->findAll(),
             'posts' => $postRepository->findApproved(20),
         ]);
     }
@@ -35,6 +36,11 @@ class PostController extends AbstractController
      */
     public function new(Request $request, FileUploader $fileUploader): Response
     {
+        if(!$this->getUser()){
+            $this->addFlash('error', 'Když nejsi přihlášen, nemůžeš přidávat příspěvky !');
+            return $this->redirectToRoute('post_index');
+        }
+
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
@@ -52,6 +58,8 @@ class PostController extends AbstractController
             $post->setAuthor($this->getUser());
             $entityManager->persist($post);
             $entityManager->flush();
+
+            $this->addFlash('success', 'Článek byl úspěšně vytvořen.');
 
             return $this->redirectToRoute('post_index');
         }
@@ -73,16 +81,15 @@ class PostController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             if(!$this->getUser()){
                 $this->addFlash('error', 'Komentář nemohl být přidán, protože nejsi přihlášen !');
-            }else{
+                return $this->redirectToRoute('post_show', ['id' => $post->getId(),]);
+            }
                 $entityManager = $this->getDoctrine()->getManager();
-
                 $postComment->setAuthor($this->getUser());
                 $postComment->setPost($post);
 
                 $entityManager->persist($postComment);
                 $entityManager->flush();
-                $this->addFlash('success', 'Komentář byl úspěšně přidán.');
-            }
+                $this->addFlash('success', 'Komentář byl přidán.');
         }
 
         return $this->render('post/show.html.twig', [
@@ -96,11 +103,22 @@ class PostController extends AbstractController
      */
     public function edit(Request $request, Post $post, FileUploader $fileUploader): Response
     {
+        if(!$this->getUser()){
+            $this->addFlash('error', 'Nemůžeš upravovat příspěvek, když nejsi přihlášen !');
+            return $this->redirectToRoute('post_index');
+        }
+
+        if($post->getAuthor() !== $this->getUser() && !in_array("ROLE_ADMIN", $this->getUser()->getRoles())){
+            $this->addFlash('error', 'Nejsi autorem příspěvku !');
+            return $this->redirectToRoute('post_index');
+        }
+
         $checkedTags = new ArrayCollection();
 
         foreach($post->getTags() as $tag) {
             $checkedTags->add($tag->getId());
         }
+
 
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
@@ -119,6 +137,8 @@ class PostController extends AbstractController
             }
             $this->getDoctrine()->getManager()->flush();
 
+            $this->addFlash('success', 'Příspěvek byl uložen.');
+
             return $this->redirectToRoute('post_index');
         }
 
@@ -134,10 +154,20 @@ class PostController extends AbstractController
      */
     public function delete(Request $request, Post $post): Response
     {
+        if(!$this->getUser()){
+            $this->addFlash('error', 'Nemůžeš odebrat tento článek, protože nejsi přihlášen !');
+            return $this->redirectToRoute('post_index');
+        }
+        if(!in_array("ROLE_ADMIN", $this->getUser()->getRoles())){
+            $this->addFlash('error', 'Nemůžeš odebrat tento článek !');
+            return $this->redirectToRoute('post_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($post);
             $entityManager->flush();
+            $this->addFlash('success', 'Příspěvek byl smazán.');
         }
 
         return $this->redirectToRoute('post_index');
